# config/api-my-stack/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/stacks": {
      "post": {
        "operationId": "createStack",
        "summary": "",
        "description": "creates a stack",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stack"
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      },
      "get": {
        "operationId": "getAllStacks",
        "summary": "",
        "description": "returns all stacks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stack"
                  }
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      }
    },
    "/api/stacks/{id}": {
      "get": {
        "operationId": "getStackById",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stack"
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      },
      "put": {
        "operationId": "updateStack",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stack"
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      },
      "delete": {
        "operationId": "deleteStack",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["stacks"]
      }
    },
    "/api/stacks/{stackId}/technologies": {
      "post": {
        "operationId": "createTechnologyForStack",
        "summary": "",
        "description": "creates a technology for a specific stack",
        "parameters": [
          {
            "name": "stackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStackTechnologyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StackTechnology"
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      },
      "get": {
        "operationId": "getAllTechnologiesOfStack",
        "summary": "",
        "description": "returns all technologies of a specific stack",
        "parameters": [
          {
            "name": "stackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StackTechnology"
                  }
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      }
    },
    "/api/stacks/{stackId}/technologies/{id}": {
      "get": {
        "operationId": "getTechnologyOfStackById",
        "summary": "",
        "parameters": [
          {
            "name": "stackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StackTechnology"
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      },
      "put": {
        "operationId": "updateTechnologyOfStackById",
        "summary": "",
        "parameters": [
          {
            "name": "stackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStackTechnologyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StackTechnology"
                }
              }
            }
          }
        },
        "tags": ["stacks"]
      },
      "delete": {
        "operationId": "deleteStack",
        "summary": "",
        "parameters": [
          {
            "name": "stackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["stacks"]
      }
    }
  },
  "info": {
    "title": "My Stack",
    "description": "My Stack - Share Your Technical Skills With the World",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "stacks",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateStackDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title"
          }
        },
        "required": ["title"]
      },
      "Stack": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Id"
          },
          "title": {
            "type": "string",
            "description": "Title"
          }
        },
        "required": ["id", "title"]
      },
      "UpdateStackDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Id"
          },
          "title": {
            "type": "string",
            "description": "Title"
          }
        },
        "required": ["id", "title"]
      },
      "CreateStackTechnologyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title"
          }
        },
        "required": ["title"]
      },
      "StackTechnology": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Id"
          },
          "title": {
            "type": "string",
            "description": "Title"
          }
        },
        "required": ["id", "title"]
      },
      "UpdateStackTechnologyDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title"
          }
        },
        "required": ["title"]
      }
    }
  }
}